render: Split off macros & pipeline creation from wgpu renderer
